+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 ## Unreleased
 
+- **Updated**
+  - Added feature specific base thrall art override
+
 ## 2.10.1
 
 - Added the `mindless` trait to thralls
-
 
 ## 2.10.0